requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
cloudinary
django-cloudinary-storage
psycopg2-binary
python-dotenv
pillow
